chore: upgrade blsttc to 6.0.0
Crates in the `safe_network` workspace will be getting upgraded to use this new version of blsttc,
so this dependent crate also needs to be upgraded.

The references in the Cargo manifest were changed to the short-form style, since there doesn't seem
to be a reason to preference the long-form style unless the line length is getting large.